Split scribe.py: extract parsers.py (DB-free parse stage)
Phase-0 refactor (#306), commit 2 of 5. Behaviour-preserving.

New bartleby/ingest/parsers.py holds everything that turns an archived file
into a ParsedDocument without touching the DB:
- the per-type converters (_parse_text_document, _parse_html_sec2md,
  _parse_edgar_submission, _parse_pdf_pdfplumber, _parse_pdf_docling,
  _parse_image_file) and the _parse_document dispatcher;
- image routing (_ImageRoute, _parse_image_routes) and the shared chunk/token
  plumbing (_archive, _build_chunk_inputs, _token_count, HTML_EXTENSIONS);
- the pool scaffolding: ParseConfig / ParseRequest / ParseOutcome and the
  parse_fn/warmup the pool schedules (_parse_request, _warm_worker).

The only non-paste change: the parse converters call `embed.embed_texts(...)`
module-qualified, so the existing `bartleby.ingest.embed.embed_texts` test
patch still bites. scribe.py now imports `from bartleby.ingest import parsers`
and references parsers.* at the call sites; dead imports dropped.

Tests retargeted to the new namespace: parsers._parse_document and friends,
and the pdfplumber convert patch to bartleby.ingest.parsers.pdfplumber_pipeline.
