+++ bartleby/commands/scribe.py
@@ -29,12 +29,11 @@ from __future__ import annotations
 import hashlib
 import itertools
 import json
-import shutil
 import sys
 import threading
 import time
 from concurrent.futures import FIRST_COMPLETED, ThreadPoolExecutor, as_completed, wait
-from dataclasses import dataclass, field
+from dataclasses import dataclass
 from pathlib import Path
 from typing import Callable
 
@@ -47,30 +46,22 @@ from bartleby.config import (
 )
 from bartleby.db.chunks import ChunkInput
 from bartleby.db.connection import open_db, resolve_project_name
-from bartleby.ingest import edgar as edgar_pipeline
 from bartleby.ingest import images as image_pipeline
-from bartleby.ingest import pdfplumber as pdfplumber_pipeline
+from bartleby.ingest import parsers
 from bartleby.ingest import pool
 from bartleby.ingest import resolve
-from bartleby.ingest import sec2md as sec2md_pipeline
 from bartleby.ingest.progress import ScribeProgress, _ProgressTally
 from bartleby.ingest.chunk import (
     IMAGE_EXTENSIONS,
-    PDF_EXTENSIONS,
-    ChunkRow,
     chunk_markdown_string,
-    convert_and_chunk,
     resolve_extension,
     resolve_format_filter,
 )
 from bartleby.ingest.embed import embed_texts
-from bartleby.ingest.summarize import count_tokens, summarize
-from bartleby.ingest.text import chunk_text
+from bartleby.ingest.summarize import summarize
 from bartleby.ingest.writer import (
     MAX_INGEST_ATTEMPTS,
     ImageCaption,
-    ParsedDocument,
-    ParsedImage,
     PendingImage,
     PendingSummary,
     Writer,
@@ -89,9 +80,6 @@ from bartleby.project import get_project_dir
 from bartleby.providers import Provider
 
 
-HTML_EXTENSIONS = {".html", ".htm"}
-
-
 # -------------------- shared helpers --------------------
 
 
@@ -101,15 +89,6 @@ def _hash_file(path: Path) -> str:
         for block in iter(lambda: f.read(64 * 1024), b""):
             h.update(block)
     return h.hexdigest()
-
-
-def _archive(src: Path, archive_root: Path, file_hash: str, ext: str) -> Path:
-    dest_dir = archive_root / file_hash
-    dest_dir.mkdir(parents=True, exist_ok=True)
-    dest = dest_dir / f"{file_hash}{ext}"
-    if not dest.exists():
-        shutil.copy2(src, dest)
-    return dest
 
 
 def _collect_files(
@@ -167,104 +146,16 @@ def _collect_files(
     return sources, unidentified
 
 
-def _build_chunk_inputs(
-    rows: list[ChunkRow],
-    embeddings: list[list[float]],
-    start_index: int = 0,
-) -> list[ChunkInput]:
-    return [
-        ChunkInput(
-            text=row.text,
-            embedding=emb,
-            chunk_index=start_index + i,
-            section_heading=row.section_heading,
-            page_number=row.page_number,
-            content_type=row.content_type,
-        )
-        for i, (row, emb) in enumerate(zip(rows, embeddings))
-    ]
-
-
-def _token_count(full_text: str) -> int:
-    return count_tokens(full_text) if full_text else 0
-
-
 def _summary_chunks(text: str) -> list[ChunkInput]:
     """Chunk + embed a summary's markdown body into ChunkInputs (producer side)."""
     rows = chunk_markdown_string(text)
     if not rows:
         return []
     embeddings = embed_texts([r.text for r in rows])
-    return _build_chunk_inputs(rows, embeddings)
+    return parsers._build_chunk_inputs(rows, embeddings)
 
 
-# -------------------- image parse + caption --------------------
-
-
-class _ImageRoute:
-    """One image (embedded or page-render) ready for the pipeline."""
-
-    def __init__(
-        self,
-        bytes_: bytes,
-        page_number: int | None,
-        image_index_on_page: int,
-    ):
-        self.bytes_ = bytes_
-        self.page_number = page_number              # None for standalone files
-        self.image_index_on_page = image_index_on_page  # 0 for page-renders/standalone
-
-
-def _parse_image_routes(
-    routes: list[_ImageRoute],
-    *,
-    archive_root: Path,
-    vision_enabled: bool,
-    vision_max_dimension: int,
-    vision_min_dimension: int,
-    on_warn: Callable[[str], None] | None = None,
-) -> list[ParsedImage]:
-    """Scale + archive each route into a ParsedImage (no VLM, no DB).
-
-    Sub-minimum images are dropped here, exactly as before: VLM image processors
-    crash on sub-patch-size edges and such slivers carry no describable content.
-    With no vision provider there's nothing to caption, so routes produce no
-    rows (a warning, then skipped). Notices go to ``on_warn`` (routed to the
-    parent), never the console — this runs in a spawn worker with no Live.
-    """
-    if not routes:
-        return []
-    if not vision_enabled:
-        if on_warn is not None:
-            on_warn(
-                f"Skipping {len(routes)} image(s) — no vision provider configured."
-            )
-        return []
-    parsed: list[ParsedImage] = []
-    for route in routes:
-        prepared = image_pipeline.prepare_image(
-            route.bytes_, max_dimension=vision_max_dimension,
-        )
-        if image_pipeline.is_below_vlm_minimum(
-            prepared, min_dimension=vision_min_dimension
-        ):
-            page_str = f" (page {route.page_number})" if route.page_number else ""
-            if on_warn is not None:
-                on_warn(
-                    f"Skipping image{page_str} — {prepared.width}x{prepared.height}px "
-                    f"is below the {vision_min_dimension}px vision minimum."
-                )
-            continue
-        archived = image_pipeline.archive_image(prepared, archive_root)
-        parsed.append(ParsedImage(
-            hash=prepared.hash,
-            archive_path=archived,
-            width=prepared.width,
-            height=prepared.height,
-            page_number=route.page_number,
-            image_index_on_page=route.image_index_on_page,
-        ))
-    return parsed
+# -------------------- image caption helpers --------------------
 
 
 def _analyze_image(
@@ -305,354 +196,6 @@ def _caption_from_analysis(
         analysis_json=analysis.model_dump_json(),
         analysis_model=vision_model,
         chunks=image_pipeline.analysis_to_chunk_inputs(analysis),
-    )
-
-
-# -------------------- per-type parse (no DB writes) --------------------
-
-
-def _parse_text_document(
-    archived: Path,
-    *,
-    file_hash: str,
-    file_name: str,
-    on_stage: Callable[[str], None] | None = None,
-) -> ParsedDocument:
-    """txt / html / md via the existing convert_and_chunk path."""
-    if on_stage is not None:
-        on_stage("extracting")
-    result = convert_and_chunk(archived)
-    chunks: list[ChunkInput] = []
-    if result.chunks:
-        embeddings = embed_texts([row.text for row in result.chunks])
-        chunks = _build_chunk_inputs(result.chunks, embeddings)
-    return ParsedDocument(
-        file_hash=file_hash, file_name=file_name, archive_path=archived,
-        page_count=result.page_count, token_count=_token_count(result.full_text),
-        document_chunks=chunks, images=[],
-    )
-
-
-def _sec2md_chunk_to_row(chunk, *, fallback_heading: str | None = None) -> ChunkRow:
-    """Map a ``Sec2mdChunk`` to a ``ChunkRow``, falling back to ``fallback_heading``
-    when sec2md didn't detect a section header for the chunk."""
-    return ChunkRow(
-        text=chunk.text,
-        section_heading=chunk.section_heading or fallback_heading,
-        content_type=chunk.content_type,
-        page_number=chunk.page_number,
-    )
-
-
-def _parse_html_sec2md(
-    archived: Path,
-    *,
-    file_hash: str,
-    file_name: str,
-    on_stage: Callable[[str], None] | None = None,
-) -> ParsedDocument:
-    """iXBRL EDGAR filing via sec2md."""
-    if on_stage is not None:
-        on_stage("extracting")
-    result = sec2md_pipeline.convert(archived)
-    chunks: list[ChunkInput] = []
-    if result.chunks:
-        if on_stage is not None:
-            on_stage("embedding")
-        rows = [_sec2md_chunk_to_row(c) for c in result.chunks]
-        embeddings = embed_texts([r.text for r in rows])
-        chunks = _build_chunk_inputs(rows, embeddings)
-    return ParsedDocument(
-        file_hash=file_hash, file_name=file_name, archive_path=archived,
-        page_count=result.page_count, token_count=_token_count(result.full_text),
-        document_chunks=chunks, images=[],
-    )
-
-
-def _parse_edgar_submission(
-    archived: Path,
-    *,
-    file_hash: str,
-    file_name: str,
-    on_stage: Callable[[str], None] | None = None,
-    on_warn: Callable[[str], None] | None = None,
-) -> ParsedDocument:
-    """EDGAR full-submission `.txt`: unwrap the SGML envelope and route each
-    inner document to its converter, landing as one document row.
-
-    Inner HTML/iXBRL bodies always go through sec2md (the only thing that reads
-    SEC HTML — there is no working fallback here, so the dependency is hard);
-    plain-text bodies go through the character chunker; graphics and XBRL data
-    files are skipped. Origin of each chunk is recorded in ``section_heading``.
-    """
-    if on_stage is not None:
-        on_stage("extracting")
-    inner_documents = edgar_pipeline.parse(archived.read_bytes())
-
-    rows: list[ChunkRow] = []
-    full_text_parts: list[str] = []
-    for doc in inner_documents:
-        label = doc.type or doc.filename or "document"
-        kind = edgar_pipeline.classify(doc)
-        if kind == "skip":
-            if on_warn is not None:
-                on_warn(f"{file_name}: skipping inner document {label}.")
-            continue
-        if kind == "html":
-            result = sec2md_pipeline.convert_bytes(doc.text.encode("utf-8"))
-            full_text_parts.append(result.full_text)
-            rows.extend(
-                _sec2md_chunk_to_row(c, fallback_heading=label)
-                for c in result.chunks
-            )
-        else:  # "text"
-            full_text_parts.append(doc.text)
-            rows.extend(
-                ChunkRow(text=t, section_heading=label, content_type=None)
-                for t in chunk_text(doc.text)
-            )
-
-    full_text = "\n\f\n".join(part for part in full_text_parts if part)
-    chunks: list[ChunkInput] = []
-    if rows:
-        if on_stage is not None:
-            on_stage("embedding")
-        embeddings = embed_texts([r.text for r in rows])
-        chunks = _build_chunk_inputs(rows, embeddings)
-    return ParsedDocument(
-        file_hash=file_hash, file_name=file_name, archive_path=archived,
-        page_count=None, token_count=_token_count(full_text),
-        document_chunks=chunks, images=[],
-    )
-
-
-def _parse_pdf_pdfplumber(
-    archived: Path,
-    *,
-    file_hash: str,
-    file_name: str,
-    archive_root: Path,
-    sparse_text_threshold: int,
-    ocr_min_confidence: int,
-    vision_enabled: bool,
-    vision_max_dimension: int,
-    vision_min_dimension: int,
-    on_stage: Callable[[str], None] | None = None,
-    on_warn: Callable[[str], None] | None = None,
-) -> ParsedDocument:
-    if on_stage is not None:
-        on_stage("extracting")
-    result = pdfplumber_pipeline.convert(
-        archived,
-        sparse_text_threshold=sparse_text_threshold,
-        ocr_min_confidence=ocr_min_confidence,
-    )
-
-    doc_rows: list[ChunkRow] = []
-    image_routes: list[_ImageRoute] = []
-    for page in result.pages:
-        if page.content_type is not None:
-            for piece in chunk_text(page.text):
-                doc_rows.append(ChunkRow(
-                    text=piece,
-                    section_heading=None,
-                    content_type=page.content_type,
-                    page_number=page.page_number,
-                ))
-        elif page.page_render_png is not None:
-            # Sparse page where OCR didn't clear the bar — fall back to VLM.
-            # The caption stage runs its own Tesseract pre-pass on the archived
-            # render (parse no longer forwards page-level OCR across the queue).
-            image_routes.append(_ImageRoute(
-                bytes_=page.page_render_png,
-                page_number=page.page_number,
-                image_index_on_page=0,
-            ))
-
-        for emb in page.embedded_images:
-            image_routes.append(_ImageRoute(
-                bytes_=emb.png_bytes,
-                page_number=page.page_number,
-                image_index_on_page=emb.image_index_on_page,
-            ))
-
-    chunks: list[ChunkInput] = []
-    if doc_rows:
-        if on_stage is not None:
-            on_stage("embedding")
-        embeddings = embed_texts([r.text for r in doc_rows])
-        chunks = _build_chunk_inputs(doc_rows, embeddings)
-
-    images = _parse_image_routes(
-        image_routes, archive_root=archive_root, vision_enabled=vision_enabled,
-        vision_max_dimension=vision_max_dimension,
-        vision_min_dimension=vision_min_dimension,
-        on_warn=on_warn,
-    )
-    return ParsedDocument(
-        file_hash=file_hash, file_name=file_name, archive_path=archived,
-        page_count=result.page_count, token_count=_token_count(result.full_text),
-        document_chunks=chunks, images=images,
-    )
-
-
-def _parse_pdf_docling(
-    archived: Path,
-    *,
-    file_hash: str,
-    file_name: str,
-    archive_root: Path,
-    vision_enabled: bool,
-    vision_max_dimension: int,
-    vision_min_dimension: int,
-    on_stage: Callable[[str], None] | None = None,
-    on_warn: Callable[[str], None] | None = None,
-) -> ParsedDocument:
-    from bartleby.ingest import docling as docling_pipeline
-
-    if on_stage is not None:
-        on_stage("extracting")
-    docling_result = docling_pipeline.convert(
-        archived, extract_images=vision_enabled,
-    )
-
-    chunks: list[ChunkInput] = []
-    if docling_result.chunks:
-        if on_stage is not None:
-            on_stage("embedding")
-        rows = [
-            ChunkRow(text=c.text, section_heading=c.section_heading,
-                     content_type=c.content_type, page_number=c.page_number)
-            for c in docling_result.chunks
-        ]
-        embeddings = embed_texts([r.text for r in rows])
-        chunks = _build_chunk_inputs(rows, embeddings)
-
-    # Embedded images come out of the same docling pass (no second parse).
-    image_routes: list[_ImageRoute] = []
-    if vision_enabled and docling_result.images:
-        image_routes = [
-            _ImageRoute(
-                bytes_=img.png_bytes,
-                page_number=img.page_number,
-                image_index_on_page=img.image_index_on_page,
-            )
-            for img in docling_result.images
-        ]
-    images = _parse_image_routes(
-        image_routes, archive_root=archive_root, vision_enabled=vision_enabled,
-        vision_max_dimension=vision_max_dimension,
-        vision_min_dimension=vision_min_dimension,
-        on_warn=on_warn,
-    )
-    return ParsedDocument(
-        file_hash=file_hash, file_name=file_name, archive_path=archived,
-        page_count=docling_result.page_count,
-        token_count=_token_count(docling_result.full_text),
-        document_chunks=chunks, images=images,
-    )
-
-
-def _parse_image_file(
-    archived: Path,
-    *,
-    file_hash: str,
-    file_name: str,
-    archive_root: Path,
-    vision_enabled: bool,
-    vision_max_dimension: int,
-    vision_min_dimension: int,
-    on_stage: Callable[[str], None] | None = None,
-    on_warn: Callable[[str], None] | None = None,
-) -> ParsedDocument:
-    """A standalone image: one route, no document chunks. The summarizer (if
-    enabled) builds its input from the image chunk once captioned. With vision
-    off the route produces no rows (warned + skipped in ``_parse_image_routes``),
-    so the document lands empty rather than stranding an uncaptionable row."""
-    if on_stage is not None:
-        on_stage("extracting")
-    route = _ImageRoute(
-        bytes_=archived.read_bytes(), page_number=None, image_index_on_page=0,
-    )
-    images = _parse_image_routes(
-        [route], archive_root=archive_root, vision_enabled=vision_enabled,
-        vision_max_dimension=vision_max_dimension,
-        vision_min_dimension=vision_min_dimension,
-        on_warn=on_warn,
-    )
-    return ParsedDocument(
-        file_hash=file_hash, file_name=file_name, archive_path=archived,
-        page_count=None, token_count=0, document_chunks=[], images=images,
-    )
-
-
-def _parse_document(
-    archived: Path,
-    ext: str,
-    *,
-    file_hash: str,
-    file_name: str,
-    pdf_converter: str,
-    html_converter: str,
-    sparse_text_threshold: int,
-    ocr_min_confidence: int,
-    vision_enabled: bool,
-    vision_max_dimension: int,
-    vision_min_dimension: int,
-    archive_root: Path,
-    on_stage: Callable[[str], None] | None = None,
-    on_warn: Callable[[str], None] | None = None,
-) -> ParsedDocument:
-    """Route an archived file to its converter and return a parsed result."""
-    if ext in IMAGE_EXTENSIONS:
-        return _parse_image_file(
-            archived, file_hash=file_hash, file_name=file_name,
-            archive_root=archive_root, vision_enabled=vision_enabled,
-            vision_max_dimension=vision_max_dimension,
-            vision_min_dimension=vision_min_dimension,
-            on_stage=on_stage, on_warn=on_warn,
-        )
-    if ext in PDF_EXTENSIONS:
-        # Reject HTML error pages saved with a `.pdf` extension before either
-        # backend dies deep in pdfminer with a cryptic "No /Root object!" — and
-        # *without* rerouting them into the HTML pipeline (they carry no
-        # document content; ingesting a portal error page pollutes the corpus).
-        pdfplumber_pipeline.reject_if_html(archived)
-        if pdf_converter == "docling":
-            return _parse_pdf_docling(
-                archived, file_hash=file_hash, file_name=file_name,
-                archive_root=archive_root, vision_enabled=vision_enabled,
-                vision_max_dimension=vision_max_dimension,
-                vision_min_dimension=vision_min_dimension,
-                on_stage=on_stage, on_warn=on_warn,
-            )
-        return _parse_pdf_pdfplumber(
-            archived, file_hash=file_hash, file_name=file_name,
-            archive_root=archive_root,
-            sparse_text_threshold=sparse_text_threshold,
-            ocr_min_confidence=ocr_min_confidence,
-            vision_enabled=vision_enabled,
-            vision_max_dimension=vision_max_dimension,
-            vision_min_dimension=vision_min_dimension,
-            on_stage=on_stage, on_warn=on_warn,
-        )
-    if edgar_pipeline.detect(archived):
-        # EDGAR full-submission SGML envelope — detected by content, so a `.txt`
-        # wrapper is caught here instead of falling to the character chunker.
-        return _parse_edgar_submission(
-            archived, file_hash=file_hash, file_name=file_name,
-            on_stage=on_stage, on_warn=on_warn,
-        )
-    if (
-        ext in HTML_EXTENSIONS
-        and html_converter == "sec2md"
-        and sec2md_pipeline.is_ixbrl(archived)
-    ):
-        return _parse_html_sec2md(
-            archived, file_hash=file_hash, file_name=file_name, on_stage=on_stage,
-        )
-    return _parse_text_document(
-        archived, file_hash=file_hash, file_name=file_name, on_stage=on_stage,
     )
 
 
@@ -913,45 +456,6 @@ def _summarize_all(
 # -------------------- parse pool: requests, config, worker fn --------------------
 
 
-@dataclass(frozen=True)
-class ParseConfig:
-    """Run-wide parse settings — identical for every document, shipped to each
-    worker once. Picklable scalars only (it crosses the spawn boundary)."""
-    pdf_converter: str
-    html_converter: str
-    sparse_text_threshold: int
-    ocr_min_confidence: int
-    vision_enabled: bool
-    vision_max_dimension: int
-    vision_min_dimension: int
-    archive_root: Path
-    timings: bool = False
-
-
-@dataclass
-class ParseRequest:
-    """One file to parse — the per-document unit the pool schedules."""
-    path: Path
-    ext: str
-    file_hash: str
-    file_name: str
-
-
-@dataclass
-class ParseOutcome:
-    """A parse result (or the failure that stood in for it), returned to the
-    main-process drain. ``parsed`` is None exactly when ``error`` is set."""
-    request: ParseRequest
-    parsed: ParsedDocument | None = None
-    parse_stages: dict[str, float] | None = None  # canonical stage→seconds (--timings)
-    error: str | None = None
-    offline: bool = False  # the error was an offline-mode block (hint the user)
-    # User-facing notices collected during parse. A spawn worker has no access
-    # to the parent's Rich Live console, so it never prints — it routes them
-    # here and the main-process drain emits them (see _parse_request).
-    warnings: list[str] = field(default_factory=list)
-
-
 @dataclass
 class _ResumeItem:
     """A document parsed by an earlier run that still has missing units."""
@@ -960,85 +464,12 @@ class _ResumeItem:
     file_hash: str
 
 
-def _parse_request(
-    request: ParseRequest,
-    config: ParseConfig,
-    report: Callable[[str], None],
-) -> ParseOutcome:
-    """Archive + parse one file into a ParsedDocument — the pool's ``parse_fn``.
-
-    DB-free, so it runs anywhere the pool puts it (a spawn worker, or inline at
-    ``max_workers <= 1``). It never raises: a parse failure comes back as an
-    ``error`` outcome so one bad file can't tear down the result stream. With
-    ``--timings`` it captures its own parse/embed wall-clock (the timer is built
-    here, in whatever process runs the parse) and returns it as data.
-
-    ``report(stage)`` is the pool's progress hook: it announces the file's
-    current stage so the main process can render this worker's lane. The single
-    ``on_stage`` the parsers already emit drives both the timer and ``report`` —
-    one callback, two consumers.
-
-    User-facing notices (skipped images, skipped inner documents) accumulate via
-    ``on_warn`` and ride back on the outcome: a spawn worker has no Live console,
-    so the main-process drain is the only place that can print them.
-    """
-    timer = timing.StageTimer() if config.timings else None
-    report("preparing")          # the lane shows the file the moment we pick it up
-
-    def on_stage(label: str) -> None:
-        if timer is not None:
-            timer.mark(label)
-        report(label)
-
-    warnings: list[str] = []
-    try:
-        parsed = _parse_document(
-            _archive(request.path, config.archive_root, request.file_hash, request.ext),
-            request.ext, file_hash=request.file_hash, file_name=request.file_name,
-            pdf_converter=config.pdf_converter, html_converter=config.html_converter,
-            sparse_text_threshold=config.sparse_text_threshold,
-            ocr_min_confidence=config.ocr_min_confidence,
-            vision_enabled=config.vision_enabled,
-            vision_max_dimension=config.vision_max_dimension,
-            vision_min_dimension=config.vision_min_dimension,
-            archive_root=config.archive_root,
-            on_stage=on_stage, on_warn=warnings.append,
-        )
-    except Exception as e:
-        from bartleby.lib.quiet import offline_blocked
-        return ParseOutcome(
-            request=request, error=str(e), offline=offline_blocked(e),
-            warnings=warnings,
-        )
-
-    if timer is not None:
-        timer.finish()
-    return ParseOutcome(
-        request=request, parsed=parsed,
-        parse_stages=dict(timer.totals) if timer is not None else None,
-        warnings=warnings,
-    )
-
-
-def _warm_worker(config: ParseConfig) -> None:
-    """Pool-initializer hook: load the models this run needs, once, up front, so
-    a worker's first document doesn't pay the load. Embeddings always; docling's
-    layout/table models only when a docling converter is active."""
-    from bartleby.ingest import embed
-
-    embed.prewarm()
-    if "docling" in (config.pdf_converter, config.html_converter):
-        from bartleby.ingest import docling as docling_pipeline
-
-        docling_pipeline.prewarm(config.vision_enabled)
-
-
 def _classify(
     writer: Writer,
     sources: list[tuple[Path, str]],
     *,
     vision_enabled: bool,
-) -> tuple[list[ParseRequest], list[_ResumeItem], list[str], list[str]]:
+) -> tuple[list[parsers.ParseRequest], list[_ResumeItem], list[str], list[str]]:
     """Bucket each source by what the parse/caption drain still needs, all from
     DB state: already parsed+captioned (skip), parsed-but-uncaptioned (resume —
     no parse), or never parsed (hand to the pool). Summaries are not considered
@@ -1051,7 +482,7 @@ def _classify(
     DB lookup can't see the in-run twin (neither is committed yet), so we track
     hashes queued this run and divert the rest here rather than let the second
     crash ``persist_parse`` (#225)."""
-    to_parse: list[ParseRequest] = []
+    to_parse: list[parsers.ParseRequest] = []
     to_resume: list[_ResumeItem] = []
     skipped: list[str] = []
     duplicates: list[str] = []
@@ -1075,7 +506,9 @@ def _classify(
             continue
         queued_hashes.add(file_hash)
         to_parse.append(
-            ParseRequest(path=path, ext=ext, file_hash=file_hash, file_name=path.name)
+            parsers.ParseRequest(
+                path=path, ext=ext, file_hash=file_hash, file_name=path.name
+            )
         )
     return to_parse, to_resume, skipped, duplicates
 
@@ -1195,7 +628,7 @@ def main(
 
     console.big(f"Ingesting {len(sources)} file(s) into project '{project_name}'")
 
-    parse_config = ParseConfig(
+    parse_config = parsers.ParseConfig(
         pdf_converter=pdf_converter_name,
         html_converter=html_converter_name,
         sparse_text_threshold=int(
@@ -1303,10 +736,10 @@ def main(
             parse_phase.start(len(to_parse))
             for outcome in pool.parse_stream(
                 to_parse,
-                parse_fn=_parse_request,
+                parse_fn=parsers._parse_request,
                 config=parse_config,
                 max_workers=max_workers,
-                warmup=_warm_worker,
+                warmup=parsers._warm_worker,
                 verbose=verbose,
                 required_models=resolve._required_hf_models(
                     pdf_converter_name, html_converter_name
